Finalized refactoring. Added documentation
