+++ CareKit/CareKit/Synchronized View Controllers/Task/OCKMultiLogTaskViewController.swift
@@ -6,21 +6,14 @@
 //
 
 import UIKit
-import CareKitUI
-import CareKitStore
 
-/// An synchronized view controller that displays a single event and it's outcomes and allows the patient to log outcomes.
+/// An synchronized view controller that displays a single event and it's outcomes and allows the patient to log outcomes with different values.
 ///
 /// - Note: `OCKEventViewController`s are created by specifying a task and an event query. If the event query
 /// returns more than one event, only the first event will be displayed.
-open class OCKMultiLogTaskViewController<Store: OCKStoreProtocol>: OCKEventViewController<Store>, OCKMultiLogTaskViewDelegate {
+open class OCKMultiLogTaskViewController<Store: OCKStoreProtocol>: OCKLogTaskViewController<Store, OCKMultiLogTaskView>, OCKMultiLogTaskViewDelegate {
 
     private var logOptions = [String]()
-    
-    public var taskView: OCKMultiLogTaskView {
-        guard let view = view as? OCKMultiLogTaskView else { fatalError("Unexpected type") }
-        return view
-    }
     
     /// Initialize using an identifier.
     ///
@@ -29,6 +22,7 @@ open class OCKMultiLogTaskViewController<Store: OCKStoreProtocol>: OCKEventViewC
     ///   - storeManager: A store manager that will be used to provide synchronization.
     ///   - taskIdentifier: The identifier event's task.
     ///   - eventQuery: An event query that specifies which events will be queried and displayed.
+    ///   - logOptions: A list of options to be displayed.
     public init(storeManager: OCKSynchronizedStoreManager<Store>, taskIdentifier: String, eventQuery: OCKEventQuery, logOptions:[String] = []) {
         super.init(storeManager: storeManager, taskIdentifier: taskIdentifier, eventQuery: eventQuery,
                    loadDefaultView: { OCKBindableMultiLogTaskView<Store.Task, Store.Outcome>() })
@@ -43,6 +37,7 @@ open class OCKMultiLogTaskViewController<Store: OCKStoreProtocol>: OCKEventViewC
     ///   - storeManager: A store manager that will be used to provide synchronization.
     ///   - task: The task to which the event to be displayed belongs.
     ///   - eventQuery: An event query that specifies which events will be queried and displayed.
+    ///   - logOptions: A list of options to be displayed.
     public convenience init(storeManager: OCKSynchronizedStoreManager<Store>, task: Store.Task, eventQuery: OCKEventQuery, logOptions:[String] = []) {
         self.init(storeManager: storeManager, taskIdentifier: task.identifier, eventQuery: eventQuery, logOptions: logOptions)
     }
@@ -53,44 +48,10 @@ open class OCKMultiLogTaskViewController<Store: OCKStoreProtocol>: OCKEventViewC
     
     override open func viewDidLoad() {
         super.viewDidLoad()
-        taskView.delegate = self
         taskView.multiLogDelegate = self
     }
     
-    // MARK: OCKSimpleLogTaskViewDelegate
-    
-    /// This method will be called each time the taps on a logged record. Override this method in a subclass to change the behavior.
-    ///
-    /// - Parameters:
-    ///   - simpleLogView: The view whose button was tapped.
-    ///   - button: The button that was tapped.
-    ///   - index: The index of the button that was tapped.
-    open func logTaskView(_ logTaskView: OCKLogTaskView, didSelectItem button: OCKButton, at index: Int) {
-        let logInfo = [button.titleLabel?.text, button.detailLabel?.text]
-            .compactMap { $0 }
-            .joined(separator: " - ")
-
-        let actionSheet = UIAlertController(title: "Log Entry", message: logInfo, preferredStyle: .actionSheet)
-        let cancel = UIAlertAction(title: OCKStyle.strings.cancel, style: .default, handler: nil)
-        
-        let delete = UIAlertAction(title: OCKStyle.strings.delete, style: .destructive) { [weak self] (action) in
-            guard let self = self else { return }
-            
-            // Sort values by date value
-            let values = self.event?.convert().outcome?.values ?? []
-            let sortedValues = values.sorted {
-                guard let date1 = $0.createdAt, let date2 = $1.createdAt else { return true }
-                return date1 < date2
-            }
-            
-            guard index < sortedValues.count else { return }
-            let intValue = sortedValues[index].integerValue
-            self.deleteOutcomeValue(intValue)
-        }
-
-        [delete, cancel].forEach { actionSheet.addAction($0) }
-        present(actionSheet, animated: true, completion: nil)
-    }
+    // MARK: OCKMultiLogTaskViewDelegate
     
     /// This method will be called each time the taps on a log button. Override this method in a subclass to change the behavior.
     ///
