+++ CareKit/CareKit/Synchronized View Controllers/Task/OCKSimpleLogTaskViewController.swift
@@ -29,19 +29,12 @@
  */
 
 import UIKit
-import CareKitUI
-import CareKitStore
 
 /// An synchronized view controller that displays a single event and it's outcomes and allows the patient to log outcomes.
 ///
 /// - Note: `OCKEventViewController`s are created by specifying a task and an event query. If the event query
 /// returns more than one event, only the first event will be displayed.
-open class OCKSimpleLogTaskViewController<Store: OCKStoreProtocol>: OCKEventViewController<Store>, OCKSimpleLogTaskViewDelegate {
-    
-    public var taskView: OCKSimpleLogTaskView {
-        guard let view = view as? OCKSimpleLogTaskView else { fatalError("Unexpected type") }
-        return view
-    }
+open class OCKSimpleLogTaskViewController<Store: OCKStoreProtocol>: OCKLogTaskViewController<Store, OCKSimpleLogTaskView> {
     
     /// Initialize using an identifier.
     ///
@@ -71,42 +64,7 @@ open class OCKSimpleLogTaskViewController<Store: OCKStoreProtocol>: OCKEventView
     
     override open func viewDidLoad() {
         super.viewDidLoad()
-        taskView.delegate = self
         taskView.logButton.addTarget(self, action: #selector(outcomeButtonPressed(_:)), for: .touchUpInside)
     }
     
-    // MARK: OCKSimpleLogTaskViewDelegate
-    
-    /// This method will be called each time the taps on a logged record. Override this method in a subclass to change the behavior.
-    ///
-    /// - Parameters:
-    ///   - simpleLogView: The view whose button was tapped.
-    ///   - button: The button that was tapped.
-    ///   - index: The index of the button that was tapped.
-    open func logTaskView(_ logTaskView: OCKLogTaskView, didSelectItem button: OCKButton, at index: Int) {
-        let logInfo = [button.titleLabel?.text, button.detailLabel?.text]
-            .compactMap { $0 }
-            .joined(separator: " - ")
-
-        let actionSheet = UIAlertController(title: "Log Entry", message: logInfo, preferredStyle: .actionSheet)
-        let cancel = UIAlertAction(title: OCKStyle.strings.cancel, style: .default, handler: nil)
-        
-        let delete = UIAlertAction(title: OCKStyle.strings.delete, style: .destructive) { [weak self] (action) in
-            guard let self = self else { return }
-            
-            // Sort values by date value
-            let values = self.event?.convert().outcome?.values ?? []
-            let sortedValues = values.sorted {
-                guard let date1 = $0.createdAt, let date2 = $1.createdAt else { return true }
-                return date1 < date2
-            }
-            
-            guard index < sortedValues.count else { return }
-            let intValue = sortedValues[index].integerValue
-            self.deleteOutcomeValue(intValue)
-        }
-
-        [delete, cancel].forEach { actionSheet.addAction($0) }
-        present(actionSheet, animated: true, completion: nil)
-    }
 }
